+++ Pandora/Paradigm/Primary/Object.hs
@@ -1,3 +1,4 @@
 module Pandora.Paradigm.Primary.Object (module Exports) where
 
+import Pandora.Paradigm.Primary.Object.Ordering as Exports
 import Pandora.Paradigm.Primary.Object.Boolean as Exports