+++ CHANGELOG.md
@@ -225,3 +225,4 @@
 * Define two experimental methods in `Bindable` class - `$>>=` and `>>=$`
 * Define `here` and `there` methods from `Wedge` to `Maybe`
 * Move `Boolean` definition to its own module
+* Move `Ordering` definition to its own module